todobackend/pyproject.toml
[project]
name = "todobackend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "pillow>=12.0.0",
    "djangorestframework-simplejwt>=5.0.0",
]
